requirements.txt
langchain>=0.1.0
langchain-core>=0.1.10
langchain-openai>=0.0.5
langgraph>=0.0.20
python-dotenv>=1.0.0
openai>=1.5.0
pydantic>=2.4.0
